StMinuitVertexFinder: Explicit type cast is not necessary

+++ StGenericVertexMaker/Minuit/StMinuitVertexFinder.cxx
@@ -368,8 +368,7 @@ void StMinuitVertexFinder::calculateRanks() {
     }
 
     //Give split vertices a lower rank...   	
-    if (mLowerSplitVtxRank &&
-        (mCTBSum > ((6000.0*(float)primV->numTracksUsedInFinder()/80)+2000)))
+    if (mLowerSplitVtxRank && mCTBSum > 6000.0*primV->numTracksUsedInFinder()/80. + 2000)
       primV->setRanking(rank_cross+rank_bemc+rank_avg_dip-3); 
     else
       primV->setRanking(rank_cross+rank_bemc+rank_avg_dip);
